requirements.txt
flask==3.0.3
pyswisseph==2.10.3.2
ephem==4.1.5
pytz==2024.1
timezonefinder==6.5.2
geopy==2.4.1
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
numpy==1.26.4
python-dateutil==2.9.0
flask-cors==4.0.1
